+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
-- `mount` field from `Raxx.Request`
-- `Raxx.BasicAuth` see the BasicAuthentication extension in `Raxx.Kit`
+- Warnings to include `raxx_logger` and `raxx_view`.
+- `mount` field from `Raxx.Request`.
+- `Raxx.BasicAuth` see the BasicAuthentication extension in `Raxx.Kit`.
 - `Raxx.Session.SignedCookie` no longer supported, to be added as extension.
 - `Raxx.SimpleClient` no longer supported.
 - `Raxx.RequestID` no longer supported.