+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.x
+## [0.18.0](https://github.com/CrowdHailer/raxx/tree/0.18.0) - 2019-02-07
 
 ### Removed